+++ lib/common/domain/check_if_app_terms_should_be_shown_usecase.dart
@@ -2,6 +2,8 @@ import 'package:zachranobed/services/auth_service.dart';
 import 'package:zachranobed/features/appConfiguration/data/get_last_app_terms_version_use_case.dart';
 
 class CheckIfAppTermsShouldBeShownUseCase {
+  // TODO: - Do not use AuthService directly - it should be contained in repository
+  // Will be resolved in [ZOB-160](https://etnetera.atlassian.net/browse/ZOB-160)
   final AuthService authService;
   final GetLastAppTermsVersionUseCase _getLastAppTermsVersionUseCase;
 
@@ -10,18 +12,15 @@ class CheckIfAppTermsShouldBeShownUseCase {
       this._getLastAppTermsVersionUseCase
   );
 
-  Future<bool> checkIfAppTermsShouldBeShown() async {
-    int? lastAppTermsVersion = await _getLastAppTermsVersionUseCase.getLastAppTerms();
-    var userData = authService.getUserData();
+  Future<bool> invoke() async {
+    final lastAppTermsVersion = await _getLastAppTermsVersionUseCase.getLastAppTerms();
+    final userData = await authService.getUserData();
+    final lastAcceptedVersion = userData?.lastAcceptedAppTermsVersion;
 
-    return userData.then((value) {
-      var lastAcceptedVersion = value?.lastAcceptedAppTermsVersion;
-
-      if (lastAcceptedVersion != null && lastAppTermsVersion != null) {
-        return lastAppTermsVersion > lastAcceptedVersion;
-      } else {
-        return true;
-      }
-    });
+    if (lastAcceptedVersion != null && lastAppTermsVersion != null) {
+      return lastAppTermsVersion > lastAcceptedVersion;
+    } else {
+      return userData != null;
+    }
   }
 }