+++ lib/features/appTerms/di/app_terms_dependency_container.dart
@@ -1,8 +1,8 @@
 import 'package:get_it/get_it.dart';
 import 'package:zachranobed/features/appConfiguration/data/get_last_app_terms_version_use_case.dart';
 import 'package:zachranobed/features/appConfiguration/data/remote_app_configuration_repository.dart';
-import 'package:zachranobed/features/app_terms/domain/set_newest_accepted_app_terms_usecase.dart';
-import 'package:zachranobed/features/app_terms/domain/update_current_user_data_use_case.dart';
+import 'package:zachranobed/features/appTerms/domain/set_newest_accepted_app_terms_usecase.dart';
+import 'package:zachranobed/features/appTerms/domain/update_current_user_data_use_case.dart';
 import 'package:zachranobed/services/auth_service.dart';
 import 'package:zachranobed/services/entity_service.dart';
 
@@ -12,7 +12,8 @@ class AppTermsDependencyContainer {
   static void setup() {
     GetIt.I.registerFactory<UpdateCurrentUserDataUseCase>(
       () => UpdateCurrentUserDataUseCase(
-        GetIt.I<EntityService>()
+          GetIt.I<EntityService>(),
+          GetIt.I<AuthService>()
       )
     );
     GetIt.I.registerFactory<SetNewestAcceptedAppTermsUseCase>(