+++ lib/routes/app_router.gr.dart
@@ -10,7 +10,7 @@
 // ignore_for_file: no_leading_underscores_for_library_prefixes
 import 'package:auto_route/auto_route.dart' as _i13;
 import 'package:flutter/material.dart' as _i14;
-import 'package:zachranobed/features/app_terms/presentation/app_terms_screen.dart'
+import 'package:zachranobed/features/appTerms/presentation/app_terms_screen.dart'
     as _i1;
 import 'package:zachranobed/features/debug/debug_screen.dart' as _i4;
 import 'package:zachranobed/features/foodboxes/domain/model/box_movement.dart'
@@ -19,6 +19,7 @@ import 'package:zachranobed/features/foodboxes/presentation/screen/box_movement_
     as _i2;
 import 'package:zachranobed/features/login/presentation/screen/login_screen.dart'
     as _i8;
+import 'package:zachranobed/features/menu/presentation/menu_screen.dart' as _i9;
 import 'package:zachranobed/features/offeredfood/domain/model/offered_food.dart'
     as _i16;
 import 'package:zachranobed/features/offeredfood/presentation/screens/donated_food_detail_screen.dart'
@@ -26,7 +27,6 @@ import 'package:zachranobed/features/offeredfood/presentation/screens/donated_fo
 import 'package:zachranobed/ui/screens/change_password_screen.dart' as _i3;
 import 'package:zachranobed/ui/screens/forgot_password_screen.dart' as _i6;
 import 'package:zachranobed/ui/screens/home_screen.dart' as _i7;
-import 'package:zachranobed/ui/screens/menu_screen.dart' as _i9;
 import 'package:zachranobed/ui/screens/offer_food_screen.dart' as _i10;
 import 'package:zachranobed/ui/screens/order_shipping_of_boxes_screen.dart'
     as _i11;