+++ CHANGELOG.md
@@ -6,6 +6,9 @@ This is a changelog for **ZachranObed** application.
 ### Added
 - **ZOB-91** Add user info item.
 - **ZOB-36** Add "Empty page" for the boxes and meal section.
+- **ZOB-48** Connected app terms link in menu screen
+- **ZOB-53** Connected partners link in menu screen
+- **ZOB-54** Connected app privacy link in menu screen
 
 ### Changed
 - **ZOB-92** Update toolbar in Profile screen.
@@ -13,13 +16,13 @@ This is a changelog for **ZachranObed** application.
 - **ZOB-114** Fetch used data from entities and entity pairs collections.
 - **ZOB-125** Change FCM tokens location in Firestore.
 - **ZOB-114** Fetch user data from entities and entity pairs collections.
-- **ZOB-125** Change FCM tokens location in Firestore.
 - **ZOB-126** Fetch data for dashboard cards.
 - **ZOB-129** Fetch data for food boxes table.
 - **ZOB-133** Fetch data for offered food.
 - **ZOB-134** Write data to deliveries collection.
 - **ZOB-135** Fetch data for box movements.
 - **ZOB-150** Update pickup config time from 35min to 45min.
+- **ZOB-155** Update notification icon.
 - **ZOB-139** Added domain logic and data connection for app terms.
 
 ### Fixed